requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pandas==1.5.3
patool==1.12
Pillow==9.4.0
pytest==7.2.1
